server/services/generator/requirements.txt
Flask==2.0.3
Flask-Cors==3.0.9
Flask-RESTful==0.3.9
gunicorn==20.0.4
torch==1.11.0
statsmodels
numpy
pandas
pystan==2.19.1.1
matrixprofile-ts
pingouin
spacy==3.2.0
scikit-learn-extra
importlib-metadata==4.13.0
tslearn
sb3-contrib==1.6.0
stable-baselines3==1.6.0